# Addon Test: Optimize internal dependencies