… test-storybooks/portable-stories-kitchen-sink/react/yarn.lock