Closes #

<!-- If your PR is related to an issue, provide the number(s) above; if it resolves multiple issues, be sure to break them up (e.g. "closes #1000, closes #1001"). -->

<!--

Thank you for contributing to Storybook! Please submit all PRs to the `next` branch unless they are specific to the current release. Storybook maintainers cherry-pick bug and documentation fixes into the `main` branch as part of the release process, so you shouldn't need to worry about this. For additional guidance: https://storybook.js.org/docs/contribute

-->

## What I did

This is an effort to fix E2E test flake that in our codebase, but turned out that addon-test was missing one bit of optimization which also contributed to flake (actually to all users) so this PR fixes that too.

## Checklist for Contributors

### Testing

<!-- Please check (put an "x" inside the "[ ]") the applicable items below to communicate how to test your changes -->

#### The changes in this PR are covered in the following automated tests:

- [ ] stories
- [ ] unit tests
- [ ] integration tests
- [ ] end-to-end tests

#### Manual testing

_This section is mandatory for all contributions. If you believe no manual test is necessary, please state so explicitly. Thanks!_

<!-- Please include the steps to test your changes here. For example:

1. Run a sandbox for template, e.g. `yarn task --task sandbox --start-from auto --template react-vite/default-ts`
2. Open Storybook in your browser
3. Access X story

-->

### Documentation

<!-- Please check (put an "x" inside the "[ ]") the applicable items below to indicate which documentation has been updated. -->

- [ ] Add or update documentation reflecting your changes
- [ ] If you are deprecating/removing a feature, make sure to update
      [MIGRATION.MD](https://github.com/storybookjs/storybook/blob/next/MIGRATION.md)

## Checklist for Maintainers

- [ ] When this PR is ready for testing, make sure to add `ci:normal`, `ci:merged` or `ci:daily` GH label to it to run a specific set of sandboxes. The particular set of sandboxes can be found in `code/lib/cli-storybook/src/sandbox-templates.ts`
- [ ] Make sure this PR contains **one** of the labels below:
   <details>
     <summary>Available labels</summary>

  - `bug`: Internal changes that fixes incorrect behavior.
  - `maintenance`: User-facing maintenance tasks.
  - `dependencies`: Upgrading (sometimes downgrading) dependencies.
  - `build`: Internal-facing build tooling & test updates. Will not show up in release changelog.
  - `cleanup`: Minor cleanup style change. Will not show up in release changelog.
  - `documentation`: Documentation **only** changes. Will not show up in release changelog.
  - `feature request`: Introducing a new feature.
  - `BREAKING CHANGE`: Changes that break compatibility in some way with current major version.
  - `other`: Changes that don't fit in the above categories.

   </details>

### 🦋 Canary release

<!-- CANARY_RELEASE_SECTION -->

This PR does not have a canary release associated. You can request a canary release of this pull request by mentioning the `@storybookjs/core` team here.

_core team members can create a canary release [here](https://github.com/storybookjs/storybook/actions/workflows/canary-release-pr.yml) or locally with `gh workflow run --repo storybookjs/storybook canary-release-pr.yml --field pr=<PR_NUMBER>`_

<!-- CANARY_RELEASE_SECTION -->

<!-- BENCHMARK_SECTION -->

| name | before | after | diff | z | % |
| ---- | ------ | ----- | ---- | - | - |
| createSize |  0 B | 0 B | 0 B | - | - |
| generateSize |  78.2 MB | 78.2 MB | 4.83 kB | **1.7** | 0% |
| initSize |  143 MB | 143 MB | 4.83 kB | **1.4** | 0% |
| diffSize |  65.2 MB | 65.2 MB | 0 B | 0.61 | 0% |
| buildSize |  6.88 MB | 6.88 MB | 0 B | 1.22 | 0% |
| buildSbAddonsSize |  1.51 MB | 1.51 MB | 0 B | 0.58 | 0% |
| buildSbCommonSize |  195 kB | 195 kB | 0 B | - | 0% |
| buildSbManagerSize |  1.9 MB | 1.9 MB | 0 B | 1.22 | 0% |
| buildSbPreviewSize |  271 kB | 271 kB | 0 B | - | 0% |
| buildStaticSize |  0 B | 0 B | 0 B | - | - |
| buildPrebuildSize |  3.88 MB | 3.88 MB | 0 B | 1.22 | 0% |
| buildPreviewSize |  3 MB | 3 MB | 0 B | - | 0% |
| testBuildSize |  0 B | 0 B | 0 B | - | - |
| testBuildSbAddonsSize |  0 B | 0 B | 0 B | - | - |
| testBuildSbCommonSize |  0 B | 0 B | 0 B | - | - |
| testBuildSbManagerSize |  0 B | 0 B | 0 B | - | - |
| testBuildSbPreviewSize |  0 B | 0 B | 0 B | - | - |
| testBuildStaticSize |  0 B | 0 B | 0 B | - | - |
| testBuildPrebuildSize |  0 B | 0 B | 0 B | - | - |
| testBuildPreviewSize |  0 B | 0 B | 0 B | - | - |



| name | before | after | diff | z | % |
| ---- | ------ | ----- | ---- | - | - |
| createTime |  7.6s | 23.8s | 16.2s | **1.4** | 🔺68.1% |
| generateTime |  23.4s | 22.7s | -716ms | 0.03 | -3.1% |
| initTime |  16.9s | 17.4s | 459ms | **1.28** | 2.6% |
| buildTime |  11s | 10.8s | -222ms | **1.52** | -2.1% |
| testBuildTime |  0ms | 0ms | 0ms | - | - |
| devPreviewResponsive |  5.3s | 5s | -261ms | **-1.27** | 🔰-5.1% |
| devManagerResponsive |  3.3s | 3.1s | -196ms | **-1.5** | 🔰-6.3% |
| devManagerHeaderVisible |  560ms | 496ms | -64ms | -1.2 | -12.9% |
| devManagerIndexVisible |  633ms | 568ms | -65ms | -0.92 | -11.4% |
| devStoryVisibleUncached |  945ms | 812ms | -133ms | **-1.47** | 🔰-16.4% |
| devStoryVisible |  627ms | 566ms | -61ms | -0.75 | -10.8% |
| devAutodocsVisible |  519ms | 432ms | -87ms | **-1.3** | 🔰-20.1% |
| devMDXVisible |  482ms | 453ms | -29ms | -0.86 | -6.4% |
| buildManagerHeaderVisible |  703ms | 506ms | -197ms | -0.93 | -38.9% |
| buildManagerIndexVisible |  725ms | 518ms | -207ms | -0.95 | -40% |
| buildStoryVisible |  704ms | 508ms | -196ms | -0.87 | -38.6% |
| buildAutodocsVisible |  566ms | 416ms | -150ms | -0.84 | -36.1% |
| buildMDXVisible |  596ms | 401ms | -195ms | -0.89 | -48.6% |

<!-- BENCHMARK_SECTION -->


<!-- greptile_comment -->

## Greptile Summary

Based on the provided files and changes, here's a concise summary of the pull request:

Optimizes internal dependencies and test configurations for the Storybook Vitest plugin, focusing on improved dependency management and test environment handling.

- Added `react-dom/test-utils` and test utilities to `optimizeDeps.include` in Vitest plugin for better dependency optimization
- Added environment variable templates in `playwright.config.ts` to simplify local E2E testing setup
- Reordered addons in test storybook configuration to maintain consistent addon loading order
- Added test stories in `AddonTest.stories.tsx` to verify browser vs CLI environment behaviors

<!-- /greptile_comment -->